requirements.txt
Flask==2.3.3
qrcode[pil]==7.4.2
pillow>=10.0.0
